Basic board drawing added
Two new operations: pressing 'f' fills the board array with random
characters and then displays the board. Pressing 'd' just displays the
current board.

The code for DrawBoard is basic, but it serves as a sort of plan. The
way that the spaces and items in those spaces are drawn comes next.