Moved drawing + window logic to Game
I'm going to make the Game struct do a lot of the management of the
state. So, calling Game_draw() will draw the whole screen, with the
board. The main loop can process input and maybe change the state of the
game, but it might be more useful to call the game's x and o moves
instead.